pyproject.toml
[project]
name = "varasto"
version = "0.1.0"
description = ""
authors = [
    {name = "Matti Luukkainen",email = "mluukkai@iki.fi"}
]
license = {text = "MIT"}
requires-python = "^3.12"
dependencies = [
    "flask (>=3.0.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.0,<4.0.0)",
    "flask-migrate (>=4.0.0,<5.0.0)",
    "flask-login (>=0.6.0,<1.0.0)",
    "flask-wtf (>=1.2.0,<2.0.0)",
    "marshmallow (>=3.20.0,<4.0.0)",
    "gunicorn (>=21.0.0,<23.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "werkzeug (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "coverage (>=7.11.0,<8.0.0)",
    "pylint (>=4.0.2,<5.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
testpaths = ["src/tests", "tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
